requirements.txt
django
psycopg2
django-jazzmin
pillow
python-dotenv
drf-yasg
djangorestframework
django-debug-toolbar
djangorestframework-simplejwt
django-filter
django-parler
django-jsoneditor
django-rosetta